More renaming openapi to snakeyaml
