+++ snakeyaml-codegen-maven-plugin/src/main/java/io/helidon/codegen/snakeyaml/SnakeYAMLMojo.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.helidon.codegen.openapi;
+package io.helidon.codegen.snakeyaml;
 
 import java.io.File;
 import java.io.FileOutputStream;
