+++ snakeyaml-codegen-maven-plugin/src/test/java/io/helidon/codegen/snakeyaml/MavenPluginHelper.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.helidon.codegen.openapi;
+package io.helidon.codegen.snakeyaml;
 
 import java.io.File;
 import java.io.IOException;
