+++ snakeyaml-codegen-maven-plugin/src/test/resources/simpleTest/pom-one-class.xml
@@ -51,7 +51,7 @@
                         <inputDirectory>implementations</inputDirectory>
                         <includes>**/*.java</includes>
                     </implementationsConfig>
-                    <outputClass>io.helidon.codegen.openapi.test.GeneratedForParser</outputClass>
+                    <outputClass>io.helidon.codegen.snakeyaml.test.GeneratedForParser</outputClass>
                 </configuration>
             </plugin>
         </plugins>
